move static methods back to Script.
