+++ src/main/java/dev/jbang/Script.java
@@ -630,24 +630,4 @@ public class Script {
 		return agentOptions;
 	}
 
-	public static List<String> collectSources(List<String> lines) {
-		List<String> sources = new ArrayList<>();
-		for (String line : lines) {
-			if (!line.startsWith(Util.SOURCES_COMMENT_PREFIX))
-				continue;
-			String[] tmp1 = line.split("[ ;,]+");
-			for (int i = 1; i < tmp1.length; ++i) {
-				sources.add(tmp1[i]);
-			}
-		}
-		return sources;
-	}
-
-	public static List<String> collectSources(String content) {
-		if (content == null) {
-			return Collections.emptyList();
-		}
-		List<String> lines = Util.getLines(null, content);
-		return collectSources(lines);
-	}
 }
