+++ src/main/java/dev/jbang/Util.java
@@ -20,6 +20,7 @@ import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
+import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.Comparator;
@@ -716,5 +717,26 @@ public class Util {
 
 	public static boolean isGistURL(String scriptURL) {
 		return scriptURL.startsWith("https://gist.github.com/");
+	}
+
+	private static List<String> collectSources(List<String> lines) {
+		List<String> sources = new ArrayList<>();
+		for (String line : lines) {
+			if (!line.startsWith(Util.SOURCES_COMMENT_PREFIX))
+				continue;
+			String[] tmp1 = line.split("[ ;,]+");
+			for (int i = 1; i < tmp1.length; ++i) {
+				sources.add(tmp1[i]);
+			}
+		}
+		return sources;
+	}
+
+	public static List<String> collectSources(String content) {
+		if (content == null) {
+			return Collections.emptyList();
+		}
+		List<String> lines = getLines(null, content);
+		return collectSources(lines);
 	}
 }
